+++ libs/eavmlib/src/esp.erl
@@ -32,8 +32,10 @@
     wakeup_cause/0,
     sleep_enable_ext0_wakeup/2,
     sleep_enable_ext1_wakeup/2,
+    nvs_fetch_binary/2,
     nvs_get_binary/1, nvs_get_binary/2, nvs_get_binary/3,
     nvs_set_binary/2, nvs_set_binary/3,
+    nvs_put_binary/3,
     nvs_erase_key/1, nvs_erase_key/2,
     nvs_erase_all/0, nvs_erase_all/1,
     nvs_reformat/0,
@@ -42,6 +44,14 @@
     rtc_slow_set_binary/1,
     freq_hz/0,
     get_mac/1
+]).
+
+-deprecated([
+    {nvs_get_binary, 1, next_version},
+    {nvs_erase_key, 1, next_version},
+    {nvs_erase_all, 0, next_version},
+    {nvs_set_binary, 2, next_version},
+    {nvs_set_binary, 3, next_version}
 ]).
 
 -type esp_reset_reason() ::
@@ -133,7 +143,23 @@ sleep_enable_ext1_wakeup(_Mask, _Mode) ->
     erlang:nif_error(undefined).
 
 %%-----------------------------------------------------------------------------
+%% @param   Namespace NVS namespace
+%% @param   Key NVS key
+%% @returns tagged tuple with binary value associated with this key in NV
+%%          storage, {error, not_found} if there is no value associated with
+%%          this key, or in general {error, Reason} for any other error.
+%% @doc     Get the binary value associated with a key, or undefined, if
+%%          there is no value associated with this key.
+%% @end
+%%-----------------------------------------------------------------------------
+-spec nvs_fetch_binary(Namespace :: atom(), Key :: atom()) ->
+    {ok, binary()} | {error, not_found} | {error, atom()}.
+nvs_fetch_binary(Namespace, Key) when is_atom(Namespace) andalso is_atom(Key) ->
+    erlang:nif_error(undefined).
+
+%%-----------------------------------------------------------------------------
 %% @doc Equivalent to nvs_get_binary(?ATOMVM_NVS_NS, Key).
+%% @deprecated Please do not use this function.
 %% @end
 %%-----------------------------------------------------------------------------
 -spec nvs_get_binary(Key :: atom()) -> binary() | undefined.
@@ -151,7 +177,11 @@ nvs_get_binary(Key) when is_atom(Key) ->
 %%-----------------------------------------------------------------------------
 -spec nvs_get_binary(Namespace :: atom(), Key :: atom()) -> binary() | undefined.
 nvs_get_binary(Namespace, Key) when is_atom(Namespace) andalso is_atom(Key) ->
-    erlang:nif_error(undefined).
+    case nvs_fetch_binary(Namespace, Key) of
+        {ok, Result} -> Result;
+        {errror, not_found} -> undefined;
+        {error, OtherError} -> throw(OtherError)
+    end.
 
 %%-----------------------------------------------------------------------------
 %% @param   Namespace NVS namespace
@@ -177,6 +207,7 @@ nvs_get_binary(Namespace, Key, Default) when
 
 %%-----------------------------------------------------------------------------
 %% @doc Equivalent to nvs_set_binary(?ATOMVM_NVS_NS, Key, Value).
+%% @deprecated Please use nvs_put_binary instead.
 %% @end
 %%-----------------------------------------------------------------------------
 -spec nvs_set_binary(Key :: atom(), Value :: binary()) -> ok.
@@ -190,10 +221,26 @@ nvs_set_binary(Key, Value) when is_atom(Key) andalso is_binary(Value) ->
 %% @returns ok
 %% @doc     Set an binary value associated with a key.  If a value exists
 %%          for the specified key, it is over-written.
+%% @deprecated Please use nvs_put_binary instead.
 %% @end
 %%-----------------------------------------------------------------------------
 -spec nvs_set_binary(Namespace :: atom(), Key :: atom(), Value :: binary()) -> ok.
 nvs_set_binary(Namespace, Key, Value) when
+    is_atom(Namespace) andalso is_atom(Key) andalso is_binary(Value)
+->
+    nvs_put_binary(Namespace, Key, Value).
+
+%%-----------------------------------------------------------------------------
+%% @param   Namespace NVS namespace
+%% @param   Key NVS key
+%% @param   Value binary value
+%% @returns ok
+%% @doc     Set an binary value associated with a key.  If a value exists
+%%          for the specified key, it is over-written.
+%% @end
+%%-----------------------------------------------------------------------------
+-spec nvs_put_binary(Namespace :: atom(), Key :: atom(), Value :: binary()) -> ok.
+nvs_put_binary(Namespace, Key, Value) when
     is_atom(Namespace) andalso is_atom(Key) andalso is_binary(Value)
 ->
     erlang:nif_error(undefined).
@@ -202,6 +249,7 @@ nvs_set_binary(Namespace, Key, Value) when
 %% @param   Key NVS key
 %% @returns ok
 %% @doc Equivalent to nvs_erase_key(?ATOMVM_NVS_NS, Key).
+%% @deprecated Please do not use this function.
 %% @end
 %%-----------------------------------------------------------------------------
 -spec nvs_erase_key(Key :: atom()) -> ok.
@@ -222,6 +270,7 @@ nvs_erase_key(Namespace, Key) when is_atom(Namespace) andalso is_atom(Key) ->
 
 %%-----------------------------------------------------------------------------
 %% @doc Equivalent to nvs_erase_all(?ATOMVM_NVS_NS).
+%% @deprecated Please do not use this function.
 %% @end
 %%-----------------------------------------------------------------------------
 -spec nvs_erase_all() -> ok.